+++ src/App.tsx
@@ -1,47 +1,12 @@
 import { CssBaseline, ThemeProvider } from '@mui/material'
 import { SignRoutes } from './routes/routes'
 import { theme } from './styles/theme'
-import { ToastContainer, toast } from 'react-toastify'
+import { ToastContainer } from 'react-toastify'
 import 'react-toastify/dist/ReactToastify.min.css'
-import { AuthProvider, AuthContext } from './contexts/auth'
+import { AuthProvider } from './contexts/auth'
 import * as React from 'react'
 
 function App() {
-  const { Logout } = React.useContext(AuthContext)
-
-  React.useEffect(() => {
-    let time: any
-
-    window.onload = resetTimer
-    document.onmousemove = resetTimer
-    document.onkeydown = resetTimer
-
-    if (time > 1000) {
-      toast.warn(
-        'Você será desconectado por inatividade em 5 minutos, clique aqui para parar!',
-        {
-          position: 'top-right',
-          autoClose: 5000000,
-          hideProgressBar: false,
-          closeOnClick: true,
-          pauseOnHover: false,
-          draggable: true,
-          progress: undefined
-        }
-      )
-    }
-
-    function doSomething() {
-      Logout()
-    }
-
-    function resetTimer() {
-      clearTimeout(time)
-      time = setTimeout(doSomething, 5000)
-      console.log('resetou')
-    }
-  })
-
   return (
     <ThemeProvider theme={theme}>
       <CssBaseline />
