+++ src/utils/controlTimeOut.tsx
@@ -1,0 +1,3 @@
+export class controlTimeOut {
+    
+}
